+++ app/Settings/CommonSettings.php
@@ -9,6 +9,7 @@ class CommonSettings extends Settings
     public ?string $lightLogo;
     public ?string $darkLogo;
     public ?string $favicon;
+    public ?string $companyShortDescription;
     public ?string $facebook;
     public ?string $instagram;
     public ?string $linkedIn;
@@ -19,6 +20,8 @@ class CommonSettings extends Settings
     public ?string $CTAMessage;
     public ?string $CTAButtonLabel;
     public ?string $CTAButtonURL;
+    public ?string $whatsAppPhoneNumber;
+    public ?string $whatsAppMessage;
     public ?bool $enableSponsorSection;
     public ?array $sponsorImages;