+++ resources/views/auth/confirm-password.blade.php
@@ -5,7 +5,9 @@
             <div class="flex justify-center">
                 <div
                     class="max-w-[400px] w-full m-auto p-6 bg-white dark:bg-slate-900 shadow-md dark:shadow-gray-800 rounded-md">
-                    <a href="index.html"><img src="assets/images/logo-icon-64.png" class="mx-auto" alt=""></a>
+                    <a href="index.html">
+                        <x-curator-glider :media="(int) $common_settings->lightLogo" class="w-auto h-32 mx-auto" />
+                    </a>
                     <h5 class="my-6 text-xl font-semibold">Confirm Password</h5>
                     <div class="mb-4 text-sm text-gray-600 dark:text-gray-400">
                         {{ __('This is a secure area of the application. Please confirm your password before continuing.') }}