+++ resources/views/layouts/fePartials/darkMode.blade.php
@@ -1,12 +1,12 @@
 <!-- Switcher -->
 <div class="fixed top-[30%] -right-2 z-50">
     <span class="relative inline-block rotate-90">
-        <input type="checkbox" class="checkbox opacity-0 absolute" id="chk" />
+        <input type="checkbox" class="absolute opacity-0 checkbox" id="chk" />
         <label
-            class="label bg-slate-900 dark:bg-white shadow dark:shadow-gray-800 cursor-pointer rounded-full flex justify-between items-center p-1 w-14 h-8"
+            class="flex items-center justify-between h-8 p-1 rounded-full shadow cursor-pointer label bg-slate-900 dark:bg-white dark:shadow-gray-800 w-14"
             for="chk">
-            <i class="uil uil-moon text-[20px] text-yellow-500"></i>
-            <i class="uil uil-sun text-[20px] text-yellow-500"></i>
+            <i class="ri-moon-line text-[20px] text-yellow-500"></i>
+            <i class="ri-sun-line text-[20px] text-yellow-500"></i>
             <span class="ball bg-white dark:bg-slate-900 rounded-full absolute top-[2px] left-[2px] w-7 h-7"></span>
         </label>
     </span>