+++ routes/web.php
@@ -36,6 +36,7 @@ Route::group(['namespace' => 'App\Http\Controllers'], function () {
     Route::get('stripe/pay', 'StripePaymentController@index')->name('stripe.pay');
     Route::get('stripe/token', 'StripePaymentController@payment_process_3d')->name('stripe.token');
     Route::get('stripe/success', 'StripePaymentController@success')->name('stripe.success');
+    Route::get('stripe/cancel', 'StripePaymentController@cancel')->name('stripe.cancel');
 });
 
 Route::middleware(['auth:customer', 'verified'])->group(function () {