+++ Changes.md
@@ -39,7 +39,7 @@
 - Add Day in Slots
 - Lane Booking No of Players Field
 - Redirect to index page on Create & Edit
-- Date Time Picker Need to Changed
+- Date Time Picker Need to Changed - **Done**
 - Lane Booking List Sport Name - **Done**
 - Slug Need to Set Based on Title or Name
 - Image not Showing in Product Table After Selecting from Curator
@@ -69,3 +69,59 @@
     scrollbar-width: none;
 }
 ```
+
+## Frontend
+
+### Header Section
+
+- Logo Change - **Done**
+- Dynamic Nav Bar
+- Login Button - **Done** Some Functionality Pending
+
+### Footer Section
+
+- Logo Change - **Done**
+- Short About Company - **Done**
+- 2 Dynamic Links
+- Replace Newsletter with Contact Us Button with some Matter - **Done**
+- Remove After Design With - **Done**
+
+## Floating Buttons
+
+- Icon for Slide to Top - **Done**
+- Dark & Light mode - **Done**
+- WhatsApp Button Custom Message & Number - **Done**
+
+## Home Page
+
+- Count Down Section Name with App Name (Change App Name to NACL Sports Center) - **Done**
+- Products Section If Published else Coming Soon - **Done**
+
+## Lanes
+
+- Inner name Lane Rental to be Replaced by Cricket Lane Rental
+- Features Section in mobile should be at Bottom if Possible else hide
+- Booking Slot Sports Selection Centered
+- Date Selection To be Replaced
+- Min Max Player in Helper Area for No of Players
+- Light Mode Color Issue
+- Saving Data Functionality Pending
+- Need to change Lane Section with booking Lane Section to change Description and Images where Book Slot will be at left and Lane Data wil lbe at right or vice versa
+
+## Sports
+
+- No Change
+
+## Shop
+
+- Full Page show be shown only if products available - **Done**
+
+## Training Video
+
+- Testimonials still not dynamic
+
+## Register
+
+- Some Fields needs to be added - **Done**
+- Save the form to Customer Model - **Done**
+- Title Change - **Done**